+++ PAOCore/DAL/TableModels/CompanyEntity.cs
@@ -1,41 +1,37 @@
 // This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com/
 
-using PAOСore;
+using PAOCore.DAL.Models;
 
-namespace PAOCore.Models.Tables
+namespace PAOCore.DAL.TableModels
 {
     /// <summary>
     /// Таблица "Компания".
     /// </summary>
-    public class CompanyEntity: BaseEntity
+    public class CompanyEntity : BaseEntity
     {
         #region public and private fields and properties
 
         /// <summary>
         /// Название.
         /// </summary>
-        public virtual string Name { get; set; }
+        public virtual string Name { get; set; } = string.Empty;
         /// <summary>
         /// Кпп.
         /// </summary>
-        public virtual string Kpp { get; set; }
+        public virtual string Kpp { get; set; } = string.Empty;
         /// <summary>
         /// Инн.
         /// </summary>
-        public virtual string Inn { get; set; }
+        public virtual string Inn { get; set; } = string.Empty;
         /// <summary>
         /// Директор.
         /// </summary>
-        public virtual PersonEntity Director { get; set; }
-        /// <summary>
-        /// Адрес.
-        /// </summary>
-        public virtual AddressEntity Address { get; set; }
+        public virtual PersonEntity Director { get; set; } = new PersonEntity();
         /// <summary>
         /// Основания полномочий.
         /// </summary>
-        public virtual string ClientBasedOn { get; set; }
+        public virtual string ClientBasedOn { get; set; } = string.Empty;
 
         #endregion
 
@@ -48,11 +44,9 @@ namespace PAOCore.Models.Tables
                 $"{nameof(Kpp)}: {Kpp}. " +
                 $"{nameof(Inn)}: {Inn}. " +
                 $"{nameof(Director)}: {Director}. " +
-                $"{nameof(Address)}: {Address}. " +
                 $"{nameof(ClientBasedOn)}: {ClientBasedOn}. ";
         }
 
         #endregion
     }
-
 }