+++ PAOCore/DAL/TableModels/ManagerMap.cs
@@ -3,16 +3,16 @@
 
 using FluentNHibernate.Mapping;
 
-namespace PAOCore.Models.Tables
+namespace PAOCore.DAL.TableModels
 {
     public class ManagerMap : ClassMap<ManagerEntity>
     {
         public ManagerMap()
         {
-            Table("[db_scales].[MANAGER]");
+            Table("[MANAGER]");
             LazyLoad();
 
-            Id(x => x.UID).CustomSqlType("UNIQUEIDENTIFIER").Column("UID").Unique().GeneratedBy.Guid().Not.Nullable();
+            Id(x => x.Uid).CustomSqlType("UNIQUEIDENTIFIER").Column("UID").Unique().GeneratedBy.Guid().Not.Nullable();
             Map(x => x.Phone).CustomSqlType("NVARCHAR(11)").Column("PHONE").Length(11).Not.Nullable();
             References(x => x.Fio).Column("FIO_UID").Not.Nullable().Unique();
         }