+++ PAOCore/DAL/TableModels/PersonMap.cs
@@ -3,17 +3,16 @@
 
 using FluentNHibernate.Mapping;
 
-
-namespace PAOCore.Models.Tables
+namespace PAOCore.DAL.TableModels
 {
     public class PersonMap : ClassMap<PersonEntity>
     {
         public PersonMap()
         {
-            Table("[db_scales].[PERSON]");
+            Table("[PERSON]");
             LazyLoad();
 
-            Id(x => x.UID).CustomSqlType("UNIQUEIDENTIFIER").Column("UID").Unique().GeneratedBy.Guid().Not.Nullable();
+            Id(x => x.Uid).CustomSqlType("UNIQUEIDENTIFIER").Column("UID").Unique().GeneratedBy.Guid().Not.Nullable();
             Map(x => x.FirstName).CustomSqlType("NVARCHAR(20)").Column("FIRST_NAME").Length(20).Not.Nullable();
             Map(x => x.LastName).CustomSqlType("NVARCHAR(20)").Column("LAST_NAME").Length(20).Not.Nullable();
             Map(x => x.MiddleName).CustomSqlType("NVARCHAR(20)").Column("MIDDLE_NAME").Length(20).Nullable();