+++ PAOCoreTests/DAL/TableModels/AddressEntityTests.cs
@@ -1,15 +1,30 @@
 // This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
+using Microsoft.Extensions.Configuration;
 using NUnit.Framework;
-using PAOCore.Models.Tables;
+using PAOCore.DAL.Models;
+using PAOCore.DAL.TableModels;
+using PAOCore.Models;
 using PAOCoreTests;
+using System.Collections.Generic;
+using System.Linq;
 
-namespace PAOTests
+namespace PAOCoreTests.DAL.TableModels
 {
     [TestFixture]
     internal class AddressEntityTests
     {
+        #region Public and private fields and properties
+
+
+        #endregion
+
+        public AddressEntityTests()
+        {
+            TestsUtils.Setup();
+        }
+
         [Test]
         public void AddressEntity_Ctor_DoesNotThrow()
         {