+++ PAOCoreTests/DAL/TableModels/ManagerEntityTests.cs
@@ -1,24 +1,24 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-using PAOCoreTests;
 using NUnit.Framework;
-using PAOCore.Models.Tables;
+using PAOCore.DAL.TableModels;
+using PAOCoreTests;
 
-namespace PAOTests
+namespace PAOCoreTests.DAL.TableModels
 {
     [TestFixture]
-    internal class AuthorityEntityTests
+    internal class ManagerEntityTests
     {
         [Test]
-        public void AuthorityEntity_Ctor_DoesNotThrow()
+        public void ManagerEntity_Ctor_DoesNotThrow()
         {
             TestsUtils.MethodStart();
 
             Assert.DoesNotThrow(() =>
             {
-                AuthorityEntity authority = new();
-                TestContext.WriteLine($"{nameof(authority).ToUpper()}: {authority}");
+                ManagerEntity manager = new();
+                TestContext.WriteLine($"{nameof(manager).ToUpper()}: {manager}");
             });
 
             TestsUtils.MethodComplete();