+++ PAOWinForms/Properties/AssemblyInfo.cs
@@ -31,5 +31,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("0.02.030.0")]
-[assembly: AssemblyFileVersion("0.02.030.0")]
+[assembly: AssemblyVersion("0.03.010.0")]
+[assembly: AssemblyFileVersion("0.03.010.0")]